refactor(RoomView): inline useReactionActions into useMessageActions
